+++ SOD.Common/ModFolderContent/CHANGELOG.md
@@ -2,7 +2,8 @@
 **1.1.8**
 - Added non-generic bind extension on BepInEx ConfigFile object
 - Added support for enum values in interface bindings
-- Added method "UpdateConfigFileLayout" to PluginController
+- Added method "UpdateConfigFileLayout" to PluginController for IPluginBindings clean up support
+- Added Dialog Helper to create dialogs between player and npcs. (Lib.Dialog)
 
 **1.1.7**
 - Made sync disk re-raise events optional (extra parameter in builder, default true)